\name{updateGaussianWishart}
\alias{}
\title{Update Inverse Wishart distributed variances with Gaussian observations}
\description{
     }
\usage{
     }
\arguments{
       \item{y}{A vector of observations, or a matrix with each row being observations and each column a variable.
       }
     }
\details{
	\deqn{	Y_i ~ MVN( meany_i, precisionY^(-1)}
	\deqn{	precisionY ~ dwish(meanVar, shapeVar)}
}     
\seealso{
     \code{\link{}}.
     }
\examples{
     }
\keyword{}

